change the protocol to fix the deadlock
The problem is that if you write a bunch of data and there's an error,
the client keeps on trying to send data and will never see the error.

We fix this by saving errors until the end of each job.

This would be a good candidate for tokio, but I don't consider
it usable in its current state.