add exponential backoff to transaction commits
this is necessary when sometimes when a lot of transactions come in all
at once and then we run out of transaction-ids, because unfortunately
the transaction ids are generated once per second.

TODO: increase the transaction id counter to add subsecond precision.

Note that this only affects the rate that you can create transactions,
not the rate of the data itself.